+++ src/main/java/edu/illinois/library/cantaloupe/processor/codec/ImageReaderFactory.java
@@ -6,6 +6,7 @@ import edu.illinois.library.cantaloupe.processor.codec.gif.GIFImageReader;
 import edu.illinois.library.cantaloupe.processor.codec.jpeg.JPEGImageReader;
 import edu.illinois.library.cantaloupe.processor.codec.png.PNGImageReader;
 import edu.illinois.library.cantaloupe.processor.codec.tiff.TIFFImageReader;
+import edu.illinois.library.cantaloupe.processor.codec.xpm.XPMImageReader;
 import edu.illinois.library.cantaloupe.source.StreamFactory;
 import edu.illinois.library.cantaloupe.source.stream.ClosingMemoryCacheImageInputStream;
 
@@ -15,10 +16,14 @@ import java.io.InputStream;
 import java.nio.file.Path;
 import java.util.Set;
 
+/**
+ * Used for obtaining {@link ImageReader} instances.
+ */
 public final class ImageReaderFactory {
 
     private static final Set<Format> SUPPORTED_FORMATS = Set.of(
-            Format.BMP, Format.GIF, Format.JPG, Format.PNG, Format.TIF);
+            Format.BMP, Format.GIF, Format.JPG, Format.PNG, Format.TIF,
+            Format.XPM);
 
     /**
      * @return Map of available output formats for all known source formats,
@@ -39,6 +44,8 @@ public final class ImageReaderFactory {
             return new PNGImageReader();
         } else if (Format.TIF.equals(format)) {
             return new TIFFImageReader();
+        } else if (Format.XPM.equals(format)) {
+            return new XPMImageReader();
         }
         throw new IllegalArgumentException("Unsupported format: " + format);
     }
@@ -46,12 +53,12 @@ public final class ImageReaderFactory {
     /**
      * Creates a reusable instance for reading from files.
      *
-     * @param sourceFile File to read from.
      * @param format     Format of the source image.
+     * @param sourceFile File to read from.
      * @throws IllegalArgumentException if the format is unsupported.
      */
-    public ImageReader newImageReader(Path sourceFile,
-                                      Format format) throws IOException {
+    public ImageReader newImageReader(Format format,
+                                      Path sourceFile) throws IOException {
         ImageReader reader = newImageReader(format);
         reader.setSource(sourceFile);
         return reader;
@@ -60,27 +67,27 @@ public final class ImageReaderFactory {
     /**
      * <p>Creates a non-reusable instance.</p>
      *
-     * <p>{@link #newImageReader(ImageInputStream, Format)} should be preferred
+     * <p>{@link #newImageReader(Format, ImageInputStream)} should be preferred
      * when a first-class {@link ImageInputStream} can be provided.</p>
      *
-     * @param inputStream Stream to read from.
      * @param format      Format of the source image.
+     * @param inputStream Stream to read from.
      * @throws IllegalArgumentException if the format is unsupported.
      */
-    public ImageReader newImageReader(InputStream inputStream,
-                                      Format format) throws IOException {
-        return newImageReader(new ClosingMemoryCacheImageInputStream(inputStream), format);
+    public ImageReader newImageReader(Format format,
+                                      InputStream inputStream) throws IOException {
+        return newImageReader(format, new ClosingMemoryCacheImageInputStream(inputStream));
     }
 
     /**
      * Creates a non-reusable instance.
      *
-     * @param inputStream Stream to read from.
      * @param format      Format of the source image.
+     * @param inputStream Stream to read from.
      * @throws IllegalArgumentException if the format is unsupported.
      */
-    public ImageReader newImageReader(ImageInputStream inputStream,
-                                      Format format) throws IOException {
+    public ImageReader newImageReader(Format format,
+                                      ImageInputStream inputStream) throws IOException {
         ImageReader reader = newImageReader(format);
         reader.setSource(inputStream);
         return reader;
@@ -89,12 +96,12 @@ public final class ImageReaderFactory {
     /**
      * Creates a reusable instance for reading from streams.
      *
+     * @param format        Format of the source image.
      * @param streamFactory Source of stream to read from.
-     * @param format       Format of the source image.
      * @throws IllegalArgumentException if the format is unsupported.
      */
-    public ImageReader newImageReader(StreamFactory streamFactory,
-                                      Format format) throws IOException {
+    public ImageReader newImageReader(Format format,
+                                      StreamFactory streamFactory) throws IOException {
         ImageReader reader = newImageReader(format);
         reader.setSource(streamFactory);
         return reader;